# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Healthcare Appointment System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctors:
    get:
      operationId: getDoctors
      tags: [doctors]
      summary: List doctors
      parameters:
        - name: specialization
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorListResponse"
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get doctor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateDoctor
      tags: [doctors]
      summary: Update doctor profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDoctorBody"
      responses:
        "200":
          description: Updated doctor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorDetail"
  /doctors/{id}/dashboard:
    get:
      operationId: getDoctorDashboard
      tags: [doctors]
      summary: Doctor dashboard data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorDashboard"
  /doctors/{id}/analytics:
    get:
      operationId: getDoctorAnalytics
      tags: [doctors]
      summary: Doctor analytics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: range
          in: query
          schema:
            type: string
            enum: [week, month, 3months]
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorAnalytics"
  /doctors/{id}/queue:
    get:
      operationId: getDoctorQueue
      tags: [doctors]
      summary: Live queue for doctor today
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Live queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorQueue"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        errors:
          type: array
          items:
            type: string
      required:
        - success
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [patient, doctor, admin]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        patientId:
          type: ["integer", "null"]
        doctorId:
          type: ["integer", "null"]
        isActive:
          type: boolean
      required:
        - id
        - name
        - email
        - role
        - isActive
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [patient, doctor]
        inviteCode:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
        - role
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - accessToken
        - user
    Doctor:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        specialization:
          type: string
        qualification:
          type: string
        experienceYears:
          type: integer
        bio:
          type: ["string", "null"]
        consultationFee:
          type: number
        avgRating:
          type: number
        totalReviews:
          type: integer
        licenseNumber:
          type: string
        isApproved:
          type: boolean
      required:
        - id
        - userId
        - name
        - email
        - specialization
        - qualification
        - experienceYears
        - consultationFee
        - avgRating
        - totalReviews
        - licenseNumber
        - isApproved
    DoctorDetail:
      allOf:
        - $ref: "#/components/schemas/Doctor"
        - type: object
          properties:
            slots:
              type: array
              items:
                $ref: "#/components/schemas/Slot"
    DoctorListResponse:
      type: object
      properties:
        doctors:
          type: array
          items:
            $ref: "#/components/schemas/Doctor"
        total:
          type: integer
      required:
        - doctors
        - total
    UpdateDoctorBody:
      type: object
      properties:
        specialization:
          type: string
        qualification:
          type: string
        experienceYears:
          type: integer
        bio:
          type: string
        consultationFee:
          type: number
        avatarUrl:
          type: string
    DoctorDashboard:
      type: object
      properties:
        todayAppointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        queue:
          $ref: "#/components/schemas/DoctorQueue"
        stats:
          $ref: "#/components/schemas/DoctorStats"
        upcomingLeave:
          type: ["string", "null"]
      required:
        - todayAppointments
        - queue
        - stats
    DoctorStats:
      type: object
      properties:
        patientsToday:
          type: integer
        patientsThisWeek:
          type: integer
        avgRating:
          type: number
        totalCompleted:
          type: integer
      required:
        - patientsToday
        - patientsThisWeek
        - avgRating
        - totalCompleted
    DoctorQueue:
      type: object
      properties:
        currentToken:
          type: ["integer", "null"]
        nextTokens:
          type: array
          items:
            $ref: "#/components/schemas/QueueToken"
        totalRemaining:
          type: integer
      required:
        - nextTokens
        - totalRemaining
    QueueToken:
      type: object
      properties:
        tokenNumber:
          type: integer
        patientName:
          type: string
        status:
          type: string
        estimatedWaitMins:
          type: integer
      required:
        - tokenNumber
        - patientName
        - status
        - estimatedWaitMins
    DoctorAnalytics:
      type: object
      properties:
        dailyAppointments:
          type: array
          items:
            $ref: "#/components/schemas/DailyCount"
        statusDistribution:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"
        waitTimeTrend:
          type: array
          items:
            $ref: "#/components/schemas/WaitTimeTrend"
        totalAppointments:
          type: integer
        completionRate:
          type: number
      required:
        - dailyAppointments
        - statusDistribution
        - waitTimeTrend
        - totalAppointments
        - completionRate
    DailyCount:
      type: object
      properties:
        date:
          type: string
        count:
          type: integer
      required:
        - date
        - count
    StatusCount:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required:
        - status
        - count
    WaitTimeTrend:
      type: object
      properties:
        date:
          type: string
        avgWaitMins:
          type: number
      required:
        - date
        - avgWaitMins
...
    Slot:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        dayOfWeek:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        slotDurationMins:
          type: integer
        maxPatients:
          type: integer
        isActive:
          type: boolean
        availableCount:
          type: ["integer", "null"]
        bookedCount:
          type: ["integer", "null"]
      required:
        - id
        - doctorId
        - dayOfWeek
        - startTime
        - endTime
        - slotDurationMins
        - maxPatients
        - isActive
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        slotId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        tokenNumber:
          type: integer
        queuePosition:
          type: integer
        estimatedWaitMins:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, no-show]
        reason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        paymentAmount:
          type: number
        createdAt:
          type: string
        patientName:
          type: ["string", "null"]
        doctorName:
          type: ["string", "null"]
        doctorSpecialization:
          type: ["string", "null"]
      required:
        - id
        - patientId
        - doctorId
        - slotId
        - date
        - startTime
        - endTime
        - tokenNumber
        - queuePosition
        - estimatedWaitMins
        - status
        - paymentStatus
        - paymentAmount
        - createdAt
...